Move ParallaxCardStack to parallelCards package and update creator
