+++ app/src/main/java/com/naulian/composable/screens/parallelCards/ParallelCardStackUI.kt
@@ -22,8 +22,6 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.naulian.composable.component.CodeBlock
-import com.naulian.composable.screens.parallax.ParallaxCardStack
-import com.naulian.composable.screens.parallax.parallaxCardStackCode
 import com.naulian.composable.theme.ComposeTheme
 import com.naulian.modify.ExperimentalModifyApi
 import com.naulian.modify.HugeIcons
